added user route and controller
